TextElement Object

TextElement Object

Usage

TextElement(
  startIndex = NULL,
  endIndex = NULL,
  paragraphMarker = NULL,
  textRun = NULL,
  autoText = NULL
)

Arguments

startIndex

(integer) The zero-based start index of this text element, in Unicode code units.

endIndex

(integer) The zero-based end index of this text element, exclusive, in Unicode code units.

paragraphMarker

(ParagraphMarker) A marker representing the beginning of a new paragraph. The start_index and end_index of this TextElement represent the range of the paragraph. Other TextElements with an index range contained inside this paragraph's range are considered to be part of this paragraph. The range of indices of two separate paragraphs will never overlap.

textRun

(TextRun) A TextElement representing a run of text where all of the characters in the run have the same TextStyle. The start_index and end_index of TextRuns will always be fully contained in the index range of a single paragraph_marker TextElement. In other words, a TextRun will never span multiple paragraphs.

autoText

(AutoText) A TextElement representing a spot in the text that is dynamically replaced with content that can change over time.

Value

TextElement object

Details

A TextElement describes the content of a range of indices in the text content of a Shape or TableCell.