UpdateShapePropertiesRequest Object

UpdateShapePropertiesRequest Object

Usage

UpdateShapePropertiesRequest(
  objectId = NULL,
  shapeProperties = NULL,
  fields = NULL
)

Arguments

objectId

(string) The object ID of the shape the updates are applied to.

shapeProperties

(ShapeProperties) The shape properties to update.

fields

(string) The fields that should be updated. At least one field must be specified. The root shapeProperties is implied and should not be specified. A single "*" can be used as short-hand for listing every field. For example to update the shape background solid fill color, set fields to "shapeBackgroundFill.solidFill.color". To reset a property to its default value, include its field name in the field mask but leave the field itself unset.

Value

UpdateShapePropertiesRequest object

Details

Update the properties of a Shape.