Edit and view auth configuration

These functions give more control over and visibility into the auth configuration than gsd_auth() does. gsd_auth_configure() lets the user specify their own:

OAuth client, which is used when obtaining a user token.

See the vignette("get-api-credentials", package = "gargle") for more. If the user does not configure these settings, internal defaults are used.

gsd_oauth_client() retrieves the currently configured OAuth client.

Usage

gsd_auth_configure(app, path)

gsd_oauth_app()

gsd_api_key()

gsd_oauth_client()

Arguments

app

Replaced by the client argument.

path

JSON downloaded from Google Cloud Console, containing a client id and secret, in one of the forms supported for the txt argument of jsonlite::fromJSON() (typically, a file path or JSON string).

client

A Google OAuth client, presumably constructed via gargle::gargle_oauth_client_from_json(). Note, however, that it is preferred to specify the client with JSON, using the path argument.

Value

gsd_auth_configure(): An object of R6 class gargle::AuthState, invisibly.

gsd_oauth_client(): the current user-configured OAuth client.

See also

Examples

# see the current user-configured OAuth app (probaby `NULL`)
gsd_oauth_app()
#> <gargle_oauth_client>
#> name: helpful-range-376621_282b25e7c11ad30ee06e5f971f717cae
#> id: 1073903696751-gr5k1k39e97ufjl5ssklhqp9pv787rfo.apps.googleusercontent.com
#> secret: <REDACTED>
#> type: installed
#> redirect_uris: http://localhost

if (require(httr)) {
  # store current state, so we can restore
  original_app <- gsd_oauth_app()

  # bring your own app via client id (aka key) and secret
  google_app <- httr::oauth_app(
    "my-awesome-google-api-wrapping-package",
    key = "123456789.apps.googleusercontent.com",
    secret = "abcdefghijklmnopqrstuvwxyz"
  )
  gsd_auth_configure(app = google_app)

  # confirm current app
  gsd_oauth_app()

  # restore original state
  gsd_auth_configure(app = original_app)
  gsd_oauth_app()
}
#> Loading required package: httr
#> <gargle_oauth_client>
#> name: helpful-range-376621_282b25e7c11ad30ee06e5f971f717cae
#> id: 1073903696751-gr5k1k39e97ufjl5ssklhqp9pv787rfo.apps.googleusercontent.com
#> secret: <REDACTED>
#> type: installed
#> redirect_uris: http://localhost

if (FALSE) { # \dontrun{
# bring your own app via JSON downloaded from GCP Console
gsd_auth_configure(
  path = "/path/to/the/JSON/you/downloaded/from/gcp/console.json"
)
} # }