Get info on current user

Reveals the email address of the user associated with the current token. If no token has been loaded yet, this function does not initiate auth.

Usage

gsd_user()

Value

An email address or, if no token has been loaded, NULL.

See also

Examples

if (FALSE) { # \dontrun{
gsd_user()
} # }