Edit and view auth configuration

These functions give more control over and visibility into the auth configuration than slides_auth() does. slides_auth_configure() lets the user specify their own:

OAuth client, which is used when obtaining a user token.

API key. If SlidesTools is de-authorized via slides_deauth(), all requests are sent with an API key in lieu of a token.

See the vignette("get-api-credentials", package = "gargle") for more. If the user does not configure these settings, internal defaults are used.

slides_oauth_client() and slides_api_key() retrieve the currently configured OAuth client and API key, respectively.

Usage

slides_auth_configure(client, path, api_key, app)

slides_api_key()

slides_oauth_client()

Arguments

client

A Google OAuth client, presumably constructed via gargle::gargle_oauth_client_from_json(). Note, however, that it is preferred to specify the client with JSON, using the path argument.

path

JSON downloaded from Google Cloud Console, containing a client id and secret, in one of the forms supported for the txt argument of jsonlite::fromJSON() (typically, a file path or JSON string).

api_key

API key.

app

Replaced by the client argument.

Value

slides_auth_configure(): An object of R6 class gargle::AuthState, invisibly.

slides_oauth_client(): the current user-configured OAuth client.

slides_api_key(): the current user-configured API key.

See also

Examples

# see and store the current user-configured OAuth client (probaby `NULL`)
(original_client <- slides_oauth_client())
#> <gargle_oauth_client>
#> name: helpful-range-376621_282b25e7c11ad30ee06e5f971f717cae
#> id: 1073903696751-gr5k1k39e97ufjl5ssklhqp9pv787rfo.apps.googleusercontent.com
#> secret: <REDACTED>
#> type: installed
#> redirect_uris: http://localhost

# see and store the current user-configured API key (probaby `NULL`)
(original_api_key <- slides_api_key())
#> NULL

# the preferred way to configure your own client is via a JSON file
# downloaded from Google Developers Console
# this example JSON is indicative, but fake
path_to_json <- system.file(
  "extdata", "client_secret_installed.googleusercontent.com.json",
  package = "gargle"
)
slides_auth_configure(path = path_to_json)

# this is also obviously a fake API key
slides_auth_configure(api_key = "the_key_I_got_for_a_google_API")

# confirm the changes
slides_oauth_client()
#> <gargle_oauth_client>
#> name: a_project_d1c5a8066d2cbe48e8d94514dd286163
#> id: abc.apps.googleusercontent.com
#> secret: <REDACTED>
#> type: installed
#> redirect_uris: http://localhost
slides_api_key()
#> [1] "the_key_I_got_for_a_google_API"

# restore original auth config
slides_auth_configure(client = original_client, api_key = original_api_key)